LA-2373 fix integer overflow for table rows count. (#45)

+++ sql_user_check_redshift/table_rows_size_per_schema.sql
@@ -2,7 +2,7 @@ SELECT
 	svt.schema,
 	count(svt.table),
 	sum(size) AS tbl_size_in_mb,
-	sum(tbl_rows::int)
+	sum(tbl_rows)
 FROM
 	svv_table_info AS svt
 GROUP BY
